commit 04.06.2022 03:51 Refactored user activity displaying code

+++ src/main/java/com/epam/billing/entity/UserActivity.java
@@ -5,7 +5,7 @@ public class UserActivity {
     private int userActivityId;
     private int activityId;
     private int userId;
-    private int durationOfActivity;
+    private float durationOfActivity;
 
     public String toString() {
 
@@ -42,11 +42,11 @@ public class UserActivity {
         return this;
     }
 
-    public int getDurationOfActivity() {
+    public float getDurationOfActivity() {
         return durationOfActivity;
     }
 
-    public UserActivity setDurationOfActivity(int durationOfActivity) {
+    public UserActivity setDurationOfActivity(float durationOfActivity) {
         this.durationOfActivity = durationOfActivity;
         return this;
     }
